More refactoring to improve singe responsibilities
